Project Euler - Problem 0007 & 0010

+++ src/problems/p_0003/mod.rs
@@ -11,14 +11,14 @@ pub fn get_problem() -> Problem {
 }
 
 
-use crate::shared::numbers::is_prime;
+use crate::shared::math::is_prime;
 
 fn solve() -> String {
     // every iteration check if the number is prime
     // if it is, then it is the largest prime factor
     // if it is not, then divide the number by the smallest divisor (returned by is_prime().1)
 
-    let mut given_num: u128 = 600851475143;
+    let mut given_num: u64 = 600851475143;
     loop {
         let result = is_prime(given_num);
         if result.0 {
