+++ src/problems/p_0005/mod.rs
@@ -11,12 +11,12 @@ pub fn get_problem() -> Problem {
 }
 
 
-use crate::shared::numbers::lcm_multiple;
+use crate::shared::math::lcm_multiple;
 
 fn solve() -> String {
-    let mut divisors: [u128; 20] = [0; 20];
+    let mut divisors: [u64; 20] = [0; 20];
     for (i, position) in divisors.iter_mut().enumerate() {
-        *position = (i + 1) as u128;
+        *position = (i + 1) as u64;
     }
     lcm_multiple(&divisors).to_string()
 }
